Created waypoints and implemented movement controls
Created an empty object and attached the camera to it then attached the provided mouselook script to allow for camera movement.
Created waypoints for the player to click on and attached the navigation waypoint script. Added particle system to the waypoints. Gave the waypoints and fade material with the colour green.
Created a HUB canvas and added a crosshair image.
Attached the interaction script to the main camera and referenced the crosshair image so the crosshair functionality works.
Rotated the doors on the cabin so they're open and allows the player to see the outside waypoint.