Attempt 2 at tooltips
Attached the world space canvas with the tooltip image and text to an empty and gave it a box collider to see if I could interact with it. I could not and the digital tooltip didn't work.
The debug tool for the interaction script doesn't work as I cannot see the ray being drawn which means I can't find out if the empty is even interactable.
Attempted with a box object as well just to see but got the same result.